+++ multitv.customtv.php
@@ -3,11 +3,11 @@
  * multiTV
  *
  * @category 	customtv
- * @version 	1.5.4
+ * @version 	1.6
  * @license 	http://www.gnu.org/copyleft/gpl.html GNU Public License (GPL)
  * @author		Jako (thomas.jakobi@partout.info)
  *
- * @internal    description: <strong>1.5.4</strong> Transform template variables into a sortable multi item list.
+ * @internal    description: <strong>1.6</strong> Transform template variables into a sortable multi item list.
  * @internal    input option code: @INCLUDE/assets/tvs/multitv/multitv.customtv.php
  */
 if (IN_MANAGER_MODE != 'true') {
@@ -15,8 +15,8 @@ if (IN_MANAGER_MODE != 'true') {
 }
 
 // set customtv (base) path
-define(MTV_PATH, 'assets/tvs/multitv/');
-define(MTV_BASE_PATH, MODX_BASE_PATH . MTV_PATH);
+define('MTV_PATH', 'assets/tvs/multitv/');
+define('MTV_BASE_PATH', MODX_BASE_PATH . MTV_PATH);
 
 if (!class_exists('multiTV')) {
 	include MTV_BASE_PATH . 'multitv.class.php';