+++ HISTORY.md
@@ -1,5 +1,11 @@
 History
 ================================================================================
+- 1.6.1 August 16, 2013
+    - Snippet: enhanced `orderBy` parameter
+
+- 1.6 July 19, 2013
+    - CustomTV: Clipper 1.2 Compatibility
+
 - 1.5.6 July 18, 2013
     - Snippet: show multiTV value always if logged into manager regardless of document publishing status